+++ account/accountController.js
@@ -1,10 +1,14 @@
 const makeService = require("./accountAppService");
 const accountValidator = require("./accountValidator");
-const accountRepo = require("./accountRepo");
+const accountRepository = require("./accountRepository");
 const modelBuilder = require("../core/modelBuilder");
 
 module.exports = function accountModule({ app, httpRequestAdaptor }) {
-  accountService = makeService({ accountRepo, modelBuilder, accountValidator });
+  accountService = makeService({
+    accountRepository,
+    modelBuilder,
+    accountValidator
+  });
 
   app.post("/api/accounts", httpRequestAdaptor(accountService.createAccount));
   //   app.get("/api/accounts", httpRequestAdaptor(accountService.getAccounts));
